Remove wrong int casts. Big #15
git-svn-id: https://svn.code.sf.net/p/cuda-z/code/trunk@186 5bddba1b-2737-4a35-b518-cf759745d68a

+++ src/cudainfo.cu
@@ -1029,7 +1029,7 @@ static float CZCudaCalcDeviceBandwidthTestCommon (
 	cudaEventDestroy(start);
 	cudaEventDestroy(stop);
 
-	return (int)bandwidthKiBs;
+	return bandwidthKiBs;
 }
 
 /*!	\brief Run several bandwidth tests.
@@ -1468,7 +1468,7 @@ static float CZCudaCalcDevicePerformanceTest(
 	cudaEventDestroy(start);
 	cudaEventDestroy(stop);
 
-	return (int)performanceKOPs;
+	return performanceKOPs;
 }
 
 /*!	\brief Calculate performance information about CUDA-device.
